# lib/apiservers/portlayer/swagger.yml
...
swagger: "2.0"
info:
  description: "Port Layer API"
  title: "Port Layer API"
  version: "v0.0.1"
produces:
- application/json
consumes:
- application/json
schemes:
- http
paths:
...
  /storage:
    post:
      description: "Creates a location to store images"
      summary: "creates an image store"
      tags: ["storage"]
      operationId: CreateImageStore
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/ImageStore"
      responses:
        '201':
          description: "Created"
          schema:
            $ref: "#/definitions/StoreUrl"
        '409':
          description: "An image store with that name already exists."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "error"
          schema:
            $ref: "#/definitions/Error"
  /storage/{store_name}:
    get:
      description: "Retrieves a list of images given a list of image IDs, or all images in the image store if no param is passed."
      summary: "Retrieve a list of images in an image store"
      tags: ["storage"]
      operationId: ListImages
      parameters:
        - name: store_name
          type: string
          in: path
          required: true
        - name: ids
          type: string
          in: query
          type: array
          items:
            type: string
            collectionFormat: csv
      responses:
        '200':
          description: "OK"
          schema:
            type: array
            items:
              $ref: "#/definitions/Image"
        '404':
          description: "Not found"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "error"
          schema:
           $ref: "#/definitions/Error"
    post:
      description: "Creates a new image layer in an image store"
      summary: "Creates a new image layer"
      tags: ["storage"]
      operationId: WriteImage
      consumes:
        - application/octet-stream
      parameters:
        - name: image_file
          in: body
          schema:
            type: string
            format: binary
        - name: store_name
          type: string
          in: path
          required: true
        - name: image_id
          type: string
          in: query
          required: true
        - name: parent_id
          type: string
          in: query
          required: true
        - name: sum
          type: string
          in: query
          required: true
        - name: metadatakey
          type: string
          in: query
        - name: metadataval
          type: string
          in: query
      responses:
        '201':
          description: "Created"
          schema:
            $ref: "#/definitions/Image"
        default:
          description: "error"
          schema:
           $ref: "#/definitions/Error"
  /storage/{store_name}/info/{id}:
    get:
      description: "Inspect an image by id in an image store"
      summary: "Inspect an image"
      tags: ["storage"]
      operationId: GetImage
      parameters:
        - name: store_name
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true
      responses:
        '404':
          description: "Not found"
          schema:
            $ref: "#/definitions/Error"
        '200':
          description: "OK"
          schema:
            $ref: "#/definitions/Image"
        default:
          description: "error"
          schema:
            $ref: "#/definitions/Error"
  /storage/{store_name}/tar/{id}:
    get:
      description: "Get an image by id in an image store as a tar file"
      summary: "Get an image as a tar file"
      tags: ["storage"]
      operationId: GetImageTar
      parameters:
        - name: store_name
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true
      responses:
        '404':
          description: "Not found"
        '200':
          description: "OK"
          schema:
            type: string
            format: binary
        default:
          description: "error"
          schema:
            $ref: "#/definitions/Error"
  /storage/volumes/:
    get:
      description: "Get a list of available volumes"
      operationId: ListVolumes
      tags: ["storage"]
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          type: string
          in: query
          type: array
          items:
            type: string
            collectionFormat: csv
      responses:
        '200':
          description: "OK"
          schema:
            type: array
            items:
              $ref: "#/definitions/VolumeResponse"
        '404':
          description: "Not found"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "error"
          schema:
           $ref: "#/definitions/Error"
    post:
      description: "Create a volume"
      operationId: CreateVolume
      summary: "Creates a Volume with metadata that is provided from the personality"
      tags: ["storage"]
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: VolumeRequest
          required: true
          in: body
          schema:
            $ref: "#/definitions/VolumeRequest"
      responses:
        '201':
          description: "Created"
          schema:
            $ref: "#/definitions/VolumeResponse"
        '500':
          description: "Error"
          $schema:
            $ref: "#/definitions/Error"
        default:
          description: "error"
          schema:
           $ref: "#/definitions/Error"
  /storage/volumes/{name}:
    get:
      description: "Get a Volume Handle"
      operationId: GetVolume
      tags: ["storage"]
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: name
          required: true
          in: path
          type: string
      responses:
        '404':
          description: "Volume not found"
          schema:
            $ref: "#/definitions/Error"
        '200':
          description: "OK"
          schema:
            type: string
    delete:
      summary: "Remove a volume"
      tags: ["storage"]
      operationId: RemoveVolume
      parameters:
        - name: name
          type: string
          in: path
          required: true
      responses:
        '200':
          description: "Volume successfully removed"
          schema:
            $ref: "#/definitions/Error"
        '404':
          description: "Volume not found"
          schema:
            $ref: "#/definitions/Error"
        '409':
          description: "Volume in use"
          schema:
            $ref: "#/definitions/Error"
        '500':
          description: "Server Error"
          schema:
            $ref: "#/definitions/Error"
    post:
      description: "Attach a volume to a container"
      operationId: VolumeJoin
      tags: ["storage"]
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: name
          required: true
          type: string
          in: path
        - name: JoinArgs
          in: body
          required: true
          schema:
            $ref: "#/definitions/VolumeJoinConfig"
      responses:
        '200':
          description: "OK"
          schema:
            type: string
        '500':
          description: "ServerError"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "error"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Error:
    type: object
    required:
      - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
  StoreUrl:
    type: object
    required:
      - url
    properties:
      code:
        type: integer
        format: int64
      url:
        type: string
  ImageStore:
    type: object
    required:
      - name
    properties:
      name:
        type: string
  Image:
    type: object
    required:
      - ID
      - Store
    properties:
      ID:
        type: string
      SelfLink:
        type: string
      Parent:
        type: string
      Store:
        type: string
      Metadata:
        type: object
        additionalProperties:
                type: string
...
  VolumeRequest:
    type: object
    required:
      - Name
      - Store
      - Capacity
      - Driver
    properties:
      Name:
        type: string
      Driver:
        type: string
      DriverArgs:
        type: object
        additionalProperties:
          type: string
      Store:
        type: string
      Capacity:
        type: integer
        format: int64
      Metadata:
        type: object
        additionalProperties:
          type: string
  VolumeResponse:
    type: object
    required:
      - Name
      - Label
      - Store
      - Driver
    properties:
      Name:
        type: string
      Label:
        type: string
      Driver:
        type: string
      Store:
        type: string
      Metadata:
        type: object
        additionalProperties:
          type: string
  VolumeJoinConfig:
    type: object
    required:
      - Handle
      - Flags
    properties:
      Handle:
        type: string
      Flags:
        type: object
        additionalProperties:
          type: string


  # The following definitions are definitions used to get container information
...